feat(service): service id is now behind a trait
This allows the implementation of new services by appchains simply by
implementing the new `ServicId` trait (in adition to the existing
`Service`) trait.

+++ crates/client/rpc/src/providers/mod.rs
@@ -72,9 +72,9 @@ impl AddTransactionProviderGroup {
     }
 
     fn provider(&self) -> Option<&Arc<dyn AddTransactionProvider>> {
-        if self.ctx.service_status(MadaraServiceId::L2Sync as u8).is_on() {
+        if self.ctx.service_status(MadaraServiceId::L2Sync).is_on() {
             Some(&self.l2_sync)
-        } else if self.ctx.service_status(MadaraServiceId::BlockProduction as u8).is_on() {
+        } else if self.ctx.service_status(MadaraServiceId::BlockProduction).is_on() {
             Some(&self.block_production)
         } else {
             None
